move robot_data.rs to robot folder
