+++ src/lib.rs
@@ -170,7 +170,6 @@ pub mod robot;
 
 mod device_data;
 pub mod model;
-mod robot_data;
 pub mod utils;
 
 pub use exception::FrankaResult;
@@ -185,7 +184,7 @@ pub use robot::fr3::Fr3;
 pub use robot::low_pass_filter::DEFAULT_CUTOFF_FREQUENCY;
 pub use robot::low_pass_filter::MAX_CUTOFF_FREQUENCY;
 pub use robot::panda::Panda;
+pub use robot::robot_data::RobotData;
 pub use robot::robot_state::RobotState;
 pub use robot::robot_wrapper::RobotWrapper;
-pub use robot_data::RobotData;
 pub use utils::*;
