+++ src/robot/control_loop.rs
@@ -11,11 +11,11 @@ use crate::robot::control_types::{ControllerMode, ConvertMotion, RealtimeConfig,
 use crate::robot::low_pass_filter::{low_pass_filter, MAX_CUTOFF_FREQUENCY};
 use crate::robot::motion_generator_traits::MotionGeneratorTrait;
 use crate::robot::rate_limiting::{limit_rate_torques, DELTA_T, MAX_TORQUE_RATE};
+use crate::robot::robot_data::RobotData;
 use crate::robot::robot_impl::RobotImplementation;
 use crate::robot::robot_state::AbstractRobotState;
 use crate::robot::service_types::{MoveControllerMode, MoveDeviation};
 use crate::robot::types::{ControllerCommand, MotionGeneratorCommand};
-use crate::robot_data::RobotData;
 use crate::Finishable;
 
 type ControlCallback<'b, State> = &'b mut dyn FnMut(&State, &Duration) -> Torques;
